Fix: Output (-1.0, -1.0) on started, but unfinished, measurement.
Suppose a measurement period starts when the duty cycle is zero, and
ends just after a rising edge of the next PWM pulse. In this case, the
output is (100.0, 0.0), which is misleading. It should be either (0.0,
0.0) or the duty cycle and frequency of the next, unfinished, PWM
pulse.

This commit reports a failed measurement (-1.0, -1.0) if the case
above occurs, and in the more general case when a measurement has been
started but not finished.